+++ src/db.js
@@ -1,4 +1,4 @@
-import IdbSchema from 'idb-schema';
+import IdbImport from './idb-import';
 
 (function (local) {
     'use strict';
@@ -67,18 +67,12 @@ import IdbSchema from 'idb-schema';
         return key;
     }
 
-    const IndexQuery = function (table, db, indexName, preexistingError, upgradeTransaction) {
+    const IndexQuery = function (table, db, indexName, preexistingError) {
         let modifyObj = null;
 
         const runQuery = function (type, args, cursorType, direction, limitRange, filters, mapper) {
             return new Promise(function (resolve, reject) {
-                let keyRange;
-                try {
-                    keyRange = type ? IDBKeyRange[type](...args) : null;
-                } catch (e) {
-                    reject(e);
-                    return;
-                }
+                const keyRange = type ? IDBKeyRange[type](...args) : null; // May throw
                 filters = filters || [];
                 limitRange = limitRange || null;
 
@@ -86,7 +80,7 @@ import IdbSchema from 'idb-schema';
                 let counter = 0;
                 const indexArgs = [keyRange];
 
-                const transaction = upgradeTransaction || db.transaction(table, modifyObj ? transactionModes.readwrite : transactionModes.readonly);
+                const transaction = db.transaction(table, modifyObj ? transactionModes.readwrite : transactionModes.readonly);
                 transaction.onerror = e => reject(e);
                 transaction.onabort = e => reject(e);
                 transaction.oncomplete = () => resolve(results);
@@ -125,37 +119,27 @@ import IdbSchema from 'idb-schema';
                             let matchFilter = true;
                             let result = 'value' in cursor ? cursor.value : cursor.key;
 
-                            try {
+                            try { // We must manually catch for this promise as we are within an async event function
                                 filters.forEach(function (filter) {
                                     if (typeof filter[0] === 'function') {
-                                        matchFilter = matchFilter && filter[0](result);
+                                        matchFilter = matchFilter && filter[0](result); // May throw with filter on non-object
                                     } else {
-                                        matchFilter = matchFilter && (result[filter[0]] === filter[1]);
+                                        matchFilter = matchFilter && (result[filter[0]] === filter[1]); // May throw with error in filter function
                                     }
                                 });
-                            } catch (err) { // Could be filter on non-object or error in filter function
+
+                                if (matchFilter) {
+                                    counter++;
+                                    // If we're doing a modify, run it now
+                                    if (modifyObj) {
+                                        result = modifyRecord(result);  // May throw
+                                        cursor.update(result); // May throw as `result` should only be a "structured clone"-able object
+                                    }
+                                    results.push(mapper(result)); // May throw
+                                }
+                            } catch (err) {
                                 reject(err);
                                 return;
-                            }
-
-                            if (matchFilter) {
-                                counter++;
-                                // If we're doing a modify, run it now
-                                if (modifyObj) {
-                                    try {
-                                        result = modifyRecord(result);
-                                        cursor.update(result); // `result` should only be a "structured clone"-able object
-                                    } catch (err) {
-                                        reject(err);
-                                        return;
-                                    }
-                                }
-                                try {
-                                    results.push(mapper(result));
-                                } catch (err) {
-                                    reject(err);
-                                    return;
-                                }
                             }
                             cursor.continue();
                         }
@@ -183,116 +167,46 @@ import IdbSchema from 'idb-schema';
             const count = function () {
                 direction = null;
                 cursorType = 'count';
-
-                return {
-                    execute
-                };
+                return {execute};
             };
 
             const keys = function () {
                 cursorType = 'openKeyCursor';
-
-                return {
-                    desc,
-                    distinct,
-                    execute,
-                    filter,
-                    limit,
-                    map
-                };
+                return {desc, distinct, execute, filter, limit, map};
             };
 
             const limit = function (start, end) {
                 limitRange = !end ? [0, start] : [start, end];
                 error = limitRange.some(val => typeof val !== 'number') ? new Error('limit() arguments must be numeric') : error;
-
-                return {
-                    desc,
-                    distinct,
-                    filter,
-                    keys,
-                    execute,
-                    map,
-                    modify
-                };
+                return {desc, distinct, filter, keys, execute, map, modify};
             };
 
             const filter = function (prop, val) {
                 filters.push([prop, val]);
-
-                return {
-                    desc,
-                    distinct,
-                    execute,
-                    filter,
-                    keys,
-                    limit,
-                    map,
-                    modify
-                };
+                return {desc, distinct, execute, filter, keys, limit, map, modify};
             };
 
             const desc = function () {
                 direction = 'prev';
-
-                return {
-                    distinct,
-                    execute,
-                    filter,
-                    keys,
-                    limit,
-                    map,
-                    modify
-                };
+                return {distinct, execute, filter, keys, limit, map, modify};
             };
 
             const distinct = function () {
                 unique = true;
-                return {
-                    count,
-                    desc,
-                    execute,
-                    filter,
-                    keys,
-                    limit,
-                    map,
-                    modify
-                };
+                return {count, desc, execute, filter, keys, limit, map, modify};
             };
 
             const modify = function (update) {
                 modifyObj = update && typeof update === 'object' ? update : null;
-                return {
-                    execute
-                };
+                return {execute};
             };
 
             const map = function (fn) {
                 mapper = fn;
-
-                return {
-                    count,
-                    desc,
-                    distinct,
-                    execute,
-                    filter,
-                    keys,
-                    limit,
-                    modify
-                };
+                return {count, desc, distinct, execute, filter, keys, limit, modify};
             };
 
-            return {
-                count,
-                desc,
-                distinct,
-                execute,
-                filter,
-                keys,
-                limit,
-                map,
-                modify
-            };
+            return {count, desc, distinct, execute, filter, keys, limit, map, modify};
         };
 
         ['only', 'bound', 'upperBound', 'lowerBound'].forEach((name) => {
@@ -322,7 +236,7 @@ import IdbSchema from 'idb-schema';
         };
     };
 
-    const Server = function (db, name, version, noServerMethods, upgradeTransaction) {
+    const Server = function (db, name, version, noServerMethods) {
         let closed = false;
 
         this.getIndexedDB = () => db;
@@ -330,7 +244,7 @@ import IdbSchema from 'idb-schema';
 
         this.query = function (table, index) {
             const error = closed ? new Error('Database has been closed') : null;
-            return new IndexQuery(table, db, index, error, upgradeTransaction); // Does not throw by itself
+            return new IndexQuery(table, db, index, error); // Does not throw by itself
         };
 
         this.add = function (table, ...args) {
@@ -344,7 +258,7 @@ import IdbSchema from 'idb-schema';
                     return records.concat(aip);
                 }, []);
 
-                const transaction = upgradeTransaction || db.transaction(table, transactionModes.readwrite);
+                const transaction = db.transaction(table, transactionModes.readwrite);
                 transaction.onerror = e => {
                     // prevent throwing a ConstraintError and aborting (hard)
                     // https://bugzilla.mozilla.org/show_bug.cgi?id=872873
@@ -361,25 +275,15 @@ import IdbSchema from 'idb-schema';
                         key = record.key;
                         record = record.item;
                         if (key != null) {
-                            try {
-                                key = mongoifyKey(key);
-                            } catch (e) {
-                                reject(e);
-                                return true;
-                            }
+                            key = mongoifyKey(key); // May throw
                         }
                     }
 
-                    try {
-                        // Safe to add since in readwrite
-                        if (key != null) {
-                            req = store.add(record, key);
-                        } else {
-                            req = store.add(record);
-                        }
-                    } catch (e) {
-                        reject(e);
-                        return true;
+                    // Safe to add since in readwrite, but may still throw
+                    if (key != null) {
+                        req = store.add(record, key);
+                    } else {
+                        req = store.add(record);
                     }
 
                     req.onsuccess = function (e) {
@@ -414,7 +318,7 @@ import IdbSchema from 'idb-schema';
                     return records.concat(aip);
                 }, []);
 
-                const transaction = upgradeTransaction || db.transaction(table, transactionModes.readwrite);
+                const transaction = db.transaction(table, transactionModes.readwrite);
                 transaction.onerror = e => {
                     // prevent throwing aborting (hard)
                     // https://bugzilla.mozilla.org/show_bug.cgi?id=872873
@@ -432,24 +336,14 @@ import IdbSchema from 'idb-schema';
                         key = record.key;
                         record = record.item;
                         if (key != null) {
-                            try {
-                                key = mongoifyKey(key);
-                            } catch (e) {
-                                reject(e);
-                                return true;
-                            }
+                            key = mongoifyKey(key); // May throw
                         }
                     }
-                    try {
-                        // These can throw DataError, e.g., if function passed in
-                        if (key != null) {
-                            req = store.put(record, key);
-                        } else {
-                            req = store.put(record);
-                        }
-                    } catch (err) {
-                        reject(err);
-                        return true;
+                    // These can throw DataError, e.g., if function passed in
+                    if (key != null) {
+                        req = store.put(record, key);
+                    } else {
+                        req = store.put(record);
                     }
 
                     req.onsuccess = function (e) {
@@ -483,14 +377,9 @@ import IdbSchema from 'idb-schema';
                     reject(new Error('Database has been closed'));
                     return;
                 }
-                try {
-                    key = mongoifyKey(key);
-                } catch (e) {
-                    reject(e);
-                    return;
-                }
+                key = mongoifyKey(key); // May throw
 
-                const transaction = upgradeTransaction || db.transaction(table, transactionModes.readwrite);
+                const transaction = db.transaction(table, transactionModes.readwrite);
                 transaction.onerror = e => {
                     // prevent throwing and aborting (hard)
                     // https://bugzilla.mozilla.org/show_bug.cgi?id=872873
@@ -501,11 +390,7 @@ import IdbSchema from 'idb-schema';
                 transaction.oncomplete = () => resolve(key);
 
                 const store = transaction.objectStore(table);
-                try {
-                    store.delete(key);
-                } catch (err) {
-                    reject(err);
-                }
+                store.delete(key); // May throw
             });
         };
 
@@ -523,7 +408,7 @@ import IdbSchema from 'idb-schema';
                     reject(new Error('Database has been closed'));
                     return;
                 }
-                const transaction = upgradeTransaction || db.transaction(table, transactionModes.readwrite);
+                const transaction = db.transaction(table, transactionModes.readwrite);
                 transaction.onerror = e => reject(e);
                 transaction.onabort = e => reject(e);
                 transaction.oncomplete = () => resolve();
@@ -552,14 +437,9 @@ import IdbSchema from 'idb-schema';
                     reject(new Error('Database has been closed'));
                     return;
                 }
-                try {
-                    key = mongoifyKey(key);
-                } catch (e) {
-                    reject(e);
-                    return;
-                }
+                key = mongoifyKey(key); // May throw
 
-                const transaction = upgradeTransaction || db.transaction(table);
+                const transaction = db.transaction(table);
                 transaction.onerror = e => {
                     // prevent throwing and aborting (hard)
                     // https://bugzilla.mozilla.org/show_bug.cgi?id=872873
@@ -570,13 +450,7 @@ import IdbSchema from 'idb-schema';
 
                 const store = transaction.objectStore(table);
 
-                let req;
-                try {
-                    req = store.get(key);
-                } catch (err) {
-                    reject(err);
-                    return;
-                }
+                const req = store.get(key);
                 req.onsuccess = e => resolve(e.target.result);
             });
         };
@@ -587,14 +461,9 @@ import IdbSchema from 'idb-schema';
                     reject(new Error('Database has been closed'));
                     return;
                 }
-                try {
-                    key = mongoifyKey(key);
-                } catch (e) {
-                    reject(e);
-                    return;
-                }
+                key = mongoifyKey(key); // May throw
 
-                const transaction = upgradeTransaction || db.transaction(table);
+                const transaction = db.transaction(table);
                 transaction.onerror = e => {
                     // prevent throwing and aborting (hard)
                     // https://bugzilla.mozilla.org/show_bug.cgi?id=872873
@@ -605,13 +474,7 @@ import IdbSchema from 'idb-schema';
 
                 const store = transaction.objectStore(table);
 
-                let req;
-                try {
-                    req = key == null ? store.count() : store.count(key);
-                } catch (err) {
-                    reject(err);
-                    return;
-                }
+                const req = key == null ? store.count() : store.count(key); // May throw
                 req.onsuccess = e => resolve(e.target.result);
             });
         };
@@ -649,7 +512,7 @@ import IdbSchema from 'idb-schema';
         }
 
         let err;
-        [].some.call(db.objectStoreNames, storeName => {
+        Array.from(db.objectStoreNames).some(storeName => {
             if (this[storeName]) {
                 err = new Error('The store name, "' + storeName + '", which you have attempted to load, conflicts with db.js method names."');
                 this.close();
@@ -665,93 +528,10 @@ import IdbSchema from 'idb-schema';
         return err;
     };
 
-    const createSchema = function (e, request, schema, db, server, version, clearUnusedStores) {
-        if (!schema || schema.length === 0) {
-            return;
-        }
+    const open = function (db, server, version, noServerMethods) {
+        dbCache[server][version] = db;
 
-        if (clearUnusedStores) {
-            for (let i = 0; i < db.objectStoreNames.length; i++) {
-                const name = db.objectStoreNames[i];
-                if (!hasOwn.call(schema, name)) {
-                    // Errors for which we are not concerned and why:
-                    // `InvalidStateError` - We are in the upgrade transaction.
-                    // `TransactionInactiveError` (as by the upgrade having already
-                    //      completed or somehow aborting) - since we've just started and
-                    //      should be without risk in this loop
-                    // `NotFoundError` - since we are iterating the dynamically updated
-                    //      `objectStoreNames`
-                    db.deleteObjectStore(name);
-                }
-            }
-        }
-
-        let ret;
-        Object.keys(schema).some(function (tableName) {
-            const table = schema[tableName];
-            let store;
-            if (db.objectStoreNames.contains(tableName)) {
-                store = request.transaction.objectStore(tableName); // Shouldn't throw
-            } else {
-                // Errors for which we are not concerned and why:
-                // `InvalidStateError` - We are in the upgrade transaction.
-                // `ConstraintError` - We are just starting (and probably never too large anyways) for a key generator.
-                // `ConstraintError` - The above condition should prevent the name already existing.
-                //
-                // Possible errors:
-                // `TransactionInactiveError` - if the upgrade had already aborted,
-                //      e.g., from a previous `QuotaExceededError` which is supposed to nevertheless return
-                //      the store but then abort the transaction.
-                // `SyntaxError` - if an invalid `table.key.keyPath` is supplied.
-                // `InvalidAccessError` - if `table.key.autoIncrement` is `true` and `table.key.keyPath` is an
-                //      empty string or any sequence (empty or otherwise).
-                try {
-                    store = db.createObjectStore(tableName, table.key);
-                } catch (err) {
-                    ret = err;
-                    return true;
-                }
-            }
-
-            Object.keys(table.indexes || {}).some(function (indexKey) {
-                try {
-                    store.index(indexKey);
-                } catch (err) {
-                    let index = table.indexes[indexKey];
-                    index = index && typeof index === 'object' ? index : {};
-                    // Errors for which we are not concerned and why:
-                    // `InvalidStateError` - We are in the upgrade transaction and store found above should not have already been deleted.
-                    // `ConstraintError` - We have already tried getting the index, so it shouldn't already exist
-                    //
-                    // Possible errors:
-                    // `TransactionInactiveError` - if the upgrade had already aborted,
-                    //      e.g., from a previous `QuotaExceededError` which is supposed to nevertheless return
-                    //      the index object but then abort the transaction.
-                    // `SyntaxError` - If the `keyPath` (second argument) is an invalid key path
-                    // `InvalidAccessError` - If `multiEntry` on `index` is `true` and
-                    //                          `keyPath` (second argument) is a sequence
-                    try {
-                        store.createIndex(indexKey, index.keyPath || index.key || indexKey, index);
-                    } catch (err2) {
-                        ret = err2;
-                        return true;
-                    }
-                }
-            });
-        });
-        return ret;
-    };
-
-    const open = function (e, server, version, noServerMethods, upgradeTransaction) {
-        const db = e.target.result;
-
-        if (!upgradeTransaction) { // We don't want to cache, especially if `upgradeTransaction`
-                                   // fails and thus does not properly overwrite with the real
-                                   // reusable object
-            dbCache[server][version] = db;
-        }
-
-        return new Server(db, server, version, noServerMethods, upgradeTransaction);
+        return new Server(db, server, version, noServerMethods);
     };
 
     const db = {
@@ -762,124 +542,88 @@ import IdbSchema from 'idb-schema';
             const clearUnusedStores = options.clearUnusedStores !== false;
             let version = options.version || 1;
             let schema = options.schema;
+            let schemas = options.schemas;
 
             if (!dbCache[server]) {
                 dbCache[server] = {};
             }
             return new Promise(function (resolve, reject) {
                 if (dbCache[server][version]) {
-                    const s = open({
-                        target: {
-                            result: dbCache[server][version]
-                        }
-                    }, server, version, noServerMethods);
+                    const s = open(dbCache[server][version], server, version, noServerMethods);
                     if (s instanceof Error) {
                         reject(s);
                         return;
                     }
                     resolve(s);
-                } else {
-                    let idbschema;
-                    if (options.schemaBuilder) {
-                        idbschema = new IdbSchema();
-                        const _addCallback = idbschema.addCallback;
-                        idbschema.addCallback = function (cb) {
-                            function newCb (e) {
-                                cb(e, idbschema._dbjs_server);
-                            }
-                            return _addCallback.call(idbschema, newCb);
-                        };
-                        try {
-                            options.schemaBuilder(idbschema);
-                            const idbschemaVersion = idbschema.version();
-                            if (options.version && idbschemaVersion < version) {
-                                throw new Error(
-                                    'Your highest schema building (IDBSchema) version (' + idbschemaVersion + ') ' +
-                                    'must not be less than your designated version (' + version + ').'
-                                );
-                            }
-                            if (!options.version && idbschemaVersion > version) {
-                                version = idbschemaVersion;
-                            }
-                        } catch (e) {
-                            reject(e);
-                            return;
-                        }
-                    }
-
-                    if (typeof schema === 'function') {
-                        try {
-                            schema = schema();
-                        } catch (e) {
-                            reject(e);
-                            return;
-                        }
-                    }
-
-                    const request = indexedDB.open(server, version);
-                    request.onsuccess = e => {
-                        const s = open(e, server, version, noServerMethods);
-                        if (s instanceof Error) {
-                            reject(s);
-                            return;
-                        }
-                        resolve(s);
-                    };
-                    request.onerror = e => {
-                        // Prevent default for `BadVersion` and `AbortError` errors, etc.
-                        // These are not necessarily reported in console in Chrome but present; see
-                        //  https://bugzilla.mozilla.org/show_bug.cgi?id=872873
-                        //  http://stackoverflow.com/questions/36225779/aborterror-within-indexeddb-upgradeneeded-event/36266502
-                        e.preventDefault();
-                        reject(e);
-                    };
-                    request.onupgradeneeded = e => {
-                        if (idbschema) {
-                            try {
-                                const s = open(e, server, version, noServerMethods, e.target.transaction);
-                                if (s instanceof Error) {
-                                    reject(s);
-                                    return;
-                                }
-                                delete s.close; // Closing should not be done in `upgradeneeded`
-
-                                // Supplying a Promise here led to problems with Firefox which
-                                //    would lose the upgrade transaction by the time the user
-                                //    callback sought to use the Server in a (modify) query,
-                                //    perhaps due to this: http://stackoverflow.com/a/28388805/271577
-                                idbschema._dbjs_server = s;
-                                idbschema.callback()(e);
-                            } catch (idbError) {
-                                reject(idbError);
-                            }
-                            return;
-                        }
-                        const err = createSchema(e, request, schema, e.target.result, server, version, clearUnusedStores);
-                        if (err) {
-                            reject(err);
-                        }
-                    };
-                    request.onblocked = e => {
-                        const resume = new Promise(function (res, rej) {
-                            // We overwrite handlers rather than make a new
-                            //   open() since the original request is still
-                            //   open and its onsuccess will still fire if
-                            //   the user unblocks by closing the blocking
-                            //   connection
-                            request.onsuccess = (ev) => {
-                                const s = open(ev, server, version, noServerMethods);
-                                if (s instanceof Error) {
-                                    rej(s);
-                                    return;
-                                }
-                                res(s);
-                            };
-                            request.onerror = e => rej(e);
-                        });
-                        e.resume = resume;
-                        reject(e);
-                    };
+                    return;
                 }
+                const openDb = function (db) {
+                    const s = open(db, server, version, noServerMethods);
+                    if (s instanceof Error) {
+                        throw s;
+                    }
+                    return s;
+                };
+                const idbimport = new IdbImport();
+                let p = Promise.resolve();
+                if (schema || schemas || options.schemaBuilder) {
+                    const _addCallback = idbimport.addCallback;
+                    idbimport.addCallback = function (cb) {
+                        function newCb (db) {
+                            const s = open(db, server, version, noServerMethods);
+                            if (s instanceof Error) {
+                                throw s;
+                            }
+                            return cb(db, s);
+                        }
+                        return _addCallback.call(idbimport, newCb);
+                    };
+
+                    if (schema) {
+                        schemas = {[version]: schema};
+                    }
+
+                    p = p.then(() => {
+                        Object.keys(schemas || {}).forEach((schemaVersion) => {
+                            let schema = schemas[schemaVersion];
+                            if (typeof schema === 'function') {
+                                schema = schema(); // May throw
+                            }
+                            idbimport.createSchema(schema, parseInt(schemaVersion, 10), clearUnusedStores); // Could immediately throw with bad version
+                        });
+                        if (options.schemaBuilder) {
+                            return options.schemaBuilder(idbimport);
+                        }
+                    }).then(() => {
+                        const idbschemaVersion = idbimport.version();
+                        if (options.version && idbschemaVersion < version) {
+                            throw new Error(
+                                'Your highest schema building (IDBSchema) version (' + idbschemaVersion + ') ' +
+                                'must not be less than your designated version (' + version + ').'
+                            );
+                        }
+                        if (!options.version && idbschemaVersion > version) {
+                            version = idbschemaVersion;
+                        }
+                    });
+                }
+
+                p.then(() => {
+                    return idbimport.open(server, version);
+                }).catch((err) => {
+                    if (err.resume) {
+                        err.resume = err.resume.then(openDb);
+                    }
+                    if (err.retry) {
+                        const _retry = err.retry;
+                        err.retry = function () {
+                            _retry.call(err).then(openDb);
+                        };
+                    }
+                    throw err;
+                }).then(openDb).then(resolve).catch((e) => {
+                    reject(e);
+                });
             });
         },
 
@@ -890,8 +634,17 @@ import IdbSchema from 'idb-schema';
             return new Promise(function (resolve, reject) {
                 const request = indexedDB.deleteDatabase(dbName); // Does not throw
 
-                request.onsuccess = e => resolve(e);
-                request.onerror = e => reject(e); // No errors currently
+                request.onsuccess = e => {
+                    // The following is needed currently by PhantomJS (though we cannot polyfill `oldVersion`): https://github.com/ariya/phantomjs/issues/14141
+                    if (!('newVersion' in e)) {
+                        e.newVersion = null;
+                    }
+                    resolve(e);
+                };
+                request.onerror = e => { // No errors currently
+                    e.preventDefault();
+                    reject(e);
+                };
                 request.onblocked = e => {
                     // The following addresses part of https://bugzilla.mozilla.org/show_bug.cgi?id=1220279
                     e = e.newVersion === null || typeof Proxy === 'undefined' ? e : new Proxy(e, {get: function (target, name) {
@@ -915,7 +668,10 @@ import IdbSchema from 'idb-schema';
 
                             res(ev);
                         };
-                        request.onerror = e => rej(e);
+                        request.onerror = e => {
+                            e.preventDefault();
+                            rej(e);
+                        };
                     });
                     e.resume = resume;
                     reject(e);
@@ -925,11 +681,7 @@ import IdbSchema from 'idb-schema';
 
         cmp: function (param1, param2) {
             return new Promise(function (resolve, reject) {
-                try {
-                    resolve(indexedDB.cmp(param1, param2));
-                } catch (e) {
-                    reject(e);
-                }
+                resolve(indexedDB.cmp(param1, param2)); // May throw
             });
         }
     };